requirements.txt
numpy
opencv-python
pyttsx3
tensorflow
cvzone
mediapipe
pyenchant
pillow
